YMSA_Project_files/GCP/GCP/source code/requirements.txt.txt
google-cloud-language
google-cloud